% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMarkerNames.R
\name{addMarkerNames}
\alias{addMarkerNames}
\title{add marker names}
\usage{
finddups(markerdata, threshold, filename)
}
\arguments{
\item{mymatrix}{- rfmatrix}

\item{rawdata}{raw marker data file to get names from}
}
\value{
pretty rf matrix
}
\description{
function to add marker names
}

