% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRfs.R
\name{addRfs}
\alias{addRfs}
\title{Internal function to add rf and cumulative rf values to a solution object.}
\usage{
addRfs(TSPsolution, rfmat)
}
\arguments{
\item{TSPsolution}{- TSP solution}

\item{rfmat}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function
@return
  TSP solution object, containing:
    1. rf value between each marker and the next
    2. cumulative rf value}
}
\description{
Internal function to add rf and cumulative rf values to a solution object.
}

