% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStampToFilename.R
\name{addStampToFilename}
\alias{addStampToFilename}
\title{Add datestamp to file}
\usage{
addStampToFilename(filename, extension)
}
\arguments{
\item{filename}{file to get date from}

\item{extension}{file extension (ie .csv)}
}
\value{
Returns filename with date and time added
}
\description{
This function adds current time and date to filename and returns value
}

