% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClusterMST.R
\name{autoClusterMST}
\alias{autoClusterMST}
\title{Break the markers into clusters.}
\usage{
autoClusterMST(rawdata, rfmat, numChromosomes,
LKHexec, internalRfThreshold)
}
\arguments{
\item{rawdata}{- raw marker data}

\item{rfmat}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}

\item{numChromosomes}{- number of chromosomes}

\item{LKHexec}{- path and filename of LKH executable - must not contain the tilde character!}

\item{internalRfThreshold}{- if a cluster contains an rf value this large, it probably contains markers from two different chromosomes.  The clusters will be split until no rf values exceeding this threshold exist within any cluster.}
}
\value{
Returns the list of clusters of marker IDs.
}
\description{
split the MST into a number of clusters equal to 1.5 times the number of chromosomes.  A cluster, in this case, must be at least numMarkers/(2*numChromosomes) in size, so that we don't count exceedingly small groups toward the total number.
}

