% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoMergeClusters.R
\name{autoMergeClusters}
\alias{autoMergeClusters}
\title{Merge the clusters from autoClusterMST into full linkage groups.}
\usage{
autoMergeClusters(clusterlist, rfmat, LKHexec,
numChromosomes, rfCap = 0, minMergeWeight = 0.40, ...)
}
\arguments{
\item{clusterlist}{- list of clusters produced by autoClusterMST}

\item{rfmat}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}

\item{LKHexec}{- path and filename of LKH executable - must not contain the tilde character!}

\item{numChromosomes}{- number of chromosomes}

\item{rfCap}{- value at which the rf matrix was capped.}

\item{minMergeWeight}{- this is the minimum weight that will cause two clusters to be merged.  If the min weight between the two groups exceeds this threshold, the two groups will not be merged.}

\item{...}{other arguments such as rawdata}
}
\value{
Returns the list of linkage groups.
}
\description{
Starting with the clusters produced by autoClusterMST, this function merges them until we are left with full linkage groups.
}

