% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callConcorde.R
\name{callConcorde}
\alias{callConcorde}
\title{Run Concorde TSP solver.}
\usage{
callConcorde(execpath,filename, outputdir)
}
\arguments{
\item{execpath}{- path and name of Concorde executable.}

\item{filename}{path and input filename as a string}

\item{outputdir}{- directory in which Concorde will save the solution file}
}
\value{
Returns the path and filename of the Concorde solution file (.sol extension).
}
\description{
Run Concorde TSP solver.
}

