% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callLKH.R
\name{callLKH}
\alias{callLKH}
\title{Run LKH TSP solver.}
\usage{
callLKH(execpath,filename,outputdir)
}
\arguments{
\item{execpath}{- path and name of LKH executable}

\item{filename}{- path and input filename as a string}

\item{outputdir}{- directory in which LKH will save the solution file}
}
\value{
Returns the path and filename of the LKH solution file (.LKH extension).
}
\description{
Run LKH TSP solver.
}

