% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capRFmat.R
\name{capRFmat}
\alias{capRFmat}
\title{Cap the rf matrix according to the given threshold.}
\usage{
capRFmat(rfmat, threshold, capval)
}
\arguments{
\item{rfmat}{- recombination frequency matrix.}

\item{threshold}{- upper limit on meaningful rf values.}

\item{capval}{- value to which rf values above threshold will be set.}
}
\value{
matrix of recombination frequency values, with all values above threshold set to capval
}
\description{
For all recombination frequency value that are above the threshold parameter, this function raises these values to capval.  This is done in order to reduce the runtime of the TSP solvers.
}

