% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanrfmat.R
\name{cleanrfmat}
\alias{cleanrfmat}
\title{Clean up rfmat object}
\usage{
cleanrfmat(rfmat)
}
\arguments{
\item{rfmat}{- recombination frequency matrix}
}
\value{
pretty rf matrix
}
\description{
function to clean up rf matrix and make it pretty
}

