% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeHaldane.R
\name{computeHaldane}
\alias{computeHaldane}
\title{Compute genetic distances (in centiMorgans) by the Haldane method}
\usage{
computeHaldane(pathData, cutpoints)
}
\arguments{
\item{pathData}{- TSP solution object, produced by the TSPinterface.R/processConcordeOutput or TSPinterface.R/processLKHOutput functions}

\item{cutpoints}{- cutpoint array, as produced by the cutpoints.R/findcuts function.  The default is NULL, which means that pathData consists of a single chromosome and therefore cutpoints are not needed.}
}
\value{
TSP solution object identical to pathData along with genetic distances and group numbers
}
\description{
Compute genetic distances (in centiMorgans) by the Haldane method, and assign a group number to the markers based on which cutpoint group they are in.
   NOTE: Recombination frequency values should NOT be divided by 2 before being passed to this function, the function takes care of that.
}

