% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRFmat.R
\name{computeRFmat}
\alias{computeRFmat}
\title{This function computes the recombination frequency matrix.}
\usage{
computeRFmat(markerdata, corrFlag)
}
\arguments{
\item{markerdata}{- main marker data file, produced by either the readfile.R/readFile function or the duplicatemarkers.R/removedups function.}

\item{corrFlag}{- flag which controls whether or not the missing data correction method is used.  Default is FALSE.}
}
\value{
matrix of recombination frequency values
}
\description{
This function computes the recombination frequency matrix.
}

