% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMap.R
\name{createMap}
\alias{createMap}
\title{Process a linkage group with Concorde to get the final linkage map.}
\usage{
createMap(rawdata, rfmat, cluster, Concexec)
}
\arguments{
\item{rawdata}{- raw marker data}

\item{rfmat}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}

\item{cluster}{- one of the linkage groups produced by autoMergeClusters}

\item{Concexec}{- path and filename of Concorde executable - must not contain the tilde character!}
}
\value{
Returns the final linkage map.
}
\description{
Starting with the clusters produced by autoClusterMST, this function merges them until we are left with full linkage groups.
}

