% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMST.R
\name{createMST}
\alias{createMST}
\title{Create minimum spanning tree.}
\usage{
createMST(rfmat, numcuts)
}
\arguments{
\item{rfmat}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}

\item{numcuts}{- number of clusters you wish to break the MST into.}
}
\value{
List of numcuts+1 marker clusters.
}
\description{
This function creates a minimum spanning tree and breaks it into numcuts clusters.
}

