% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTSPFile.R
\name{createTSPFile}
\alias{createTSPFile}
\title{Create TSP problem file.}
\usage{
createTSPFile(markerdata,rfmatrix,strainName, dirname)
}
\arguments{
\item{markerdata}{main marker data file, produced by the readfile.R/readFile function}

\item{rfmatrix}{recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}

\item{strainName}{name of the strain as a string (no spaces), used to name the file}

\item{dirname}{directory in which to create the file}
}
\value{
Returns the path and filename of the TSP file (.tsp extension).
}
\description{
Create TSP problem file (for use with both Concorde and LKH).
 This converts the rf matrix to an integer matrix that the TSP solvers need.
 HAMILTONIAN PATH CORRECTION: This also adds a dummy node which has 0 weight to all other nodes so that the TSP solver will find a Hamiltonian path instead of a circuit.
}

