% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancedf2table.R
\name{distancedf2table}
\alias{distancedf2table}
\title{create table of marker distances and groups}
\usage{
distancedf2table(distancedf)
}
\arguments{
\item{distancedf}{dataframe containing markernames (markername) distances (`Haldane (cM)`) and groups (groupname) of markers}
}
\value{
table with chr=linkage group, pos=distance, and row.names = marker names
}
\description{
converts distancedf into table that can be converted into a map object by qtl::table2map
}

