% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcuts.R
\name{findcuts}
\alias{findcuts}
\title{Determine cutpoints in TSP solution object.}
\usage{
findcuts(tourData, numCuts, alt)
}
\arguments{
\item{tourData}{- TSP solution object, produced by the TSPinterface.R/processConcordeOutput or TSPinterface.R/processLKHOutput function}

\item{numCuts}{- number of cutpoints to find}

\item{alt}{- alternate mode, where we instead find the cutpoints that are above a given threshold (specified by the numCuts parameter).}
}
\value{
array of cutpoint info, containing:
     1. marker index of the marker at the cutpoint
     2. starting index of the group
     3. ending index of the group
     4. size of the group
}
\description{
Determine cutpoints in TSP solution object.
}

