% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGroups.R
\name{matchGroups}
\alias{matchGroups}
\title{Combine two cutpoint groups and pass them into LKH and return the value of the second largest rf value in the result.}
\usage{
matchGroups(ind_group1,ind_group2, markerData, cutList, rfmatrix, execPath)
}
\arguments{
\item{ind_group1}{- index of the first group}

\item{ind_group2}{- index of the second group}

\item{markerData}{- main marker data file, produced by either the readfile.R/readFile function or the duplicatemarkers.R/removedups function}

\item{cutList}{- cutpoint array, produced by cutpoints.R/findcuts function}

\item{rfmatrix}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}

\item{execPath}{- path to LKH executable}
}
\value{
Maximum rf value in the TSP path of the two clusters combined.
}
\description{
Combine two cutpoint groups and pass them into LKH and return the value of the second largest rf value in the result.
   This is used to pair small subgroups with a larger group by combining the small group with all larger groups and seeing which results in the smallest return value.
   The reason we take the second largest rf value in the TSP result is because the largest rf value represents the completion of the TSP circuit, so even for a single chromosome this will be a large value.
}

