% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGroups2.R
\name{matchGroups2}
\alias{matchGroups2}
\title{This is an alternate version of the matchGroups function above.  This version works with groups from the GroupList.}
\usage{
matchGroups2(group1, group2 , rawdata, rfmatrix, execPath)
}
\arguments{
\item{group1}{index of the first cluster}

\item{group2}{- index of the second cluster}

\item{rawdata}{- main marker data file, produced by either the readfile.R/readFile function or the duplicatemarkers.R/removedups function}

\item{rfmatrix}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}

\item{execPath}{- path and name of LKH executable}
}
\value{
Maximum rf value in the TSP path of the two clusters combined.
}
\description{
Combine two cutpoint groups and pass them into LKH and return the value of the second largest rf value in the result.
   This is used to pair small subgroups with a larger group by combining the small group with all larger groups and seeing which results in the smallest return value.
   The reason we take the second largest rf value in the TSP result is because the largest rf value represents the completion of the TSP circuit, so even for a single chromosome this will be a large value.
}

