% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSTgetGroup.R
\name{MSTgetGroup}
\alias{MSTgetGroup}
\title{Get the marker IDs for a particular cluster}
\usage{
MSTgetGroup(clusterinfo,clusterIndex)
}
\arguments{
\item{clusterinfo}{list of clusters, as produced by createMST function.}

\item{clusterIndex}{the index of the cluster.}
}
\value{
Marker IDs of the selected cluster.
}
\description{
This function returns the list of marker IDs for a given cluster.
}

