% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareMatrix.R
\name{pareMatrix}
\alias{pareMatrix}
\title{This function pares down a given rf matrix by keeping or removing the markers that are not in the provided subtour.}
\usage{
pareMatrix(rfmatrix, markerIndexList, keep, clusters)
}
\arguments{
\item{rfmatrix}{- matrix of recombination frequency values produced by the rfmatrix.R/computeRFmat function}

\item{markerIndexList}{- list of the indices of the markers that you want to be included in the pared down matrix.  NOTE: this list is most easily obtained by using the cutpoint array that is produced by the cutpoints.R/findcuts function.}

\item{keep}{(OPTIONAL) - this flag indicates whether you want to keep the markers that are in the markerIndexList, or discard them.}

\item{clusters}{(OPTIONAL) - this flag indicates whether you are passing in a cluster of marker ids or a markerlist.}
}
\value{
submatrix of rfmatrix that only includes or excludes the markers given in markerIndexList, depending on the value of the keep flag.
}
\description{
This function pares down a given rf matrix by keeping or removing the markers that are not in the provided subtour.
   This is used when we want to pass a subset of the markers into a TSP solver.
}

