% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSPsoln.R
\name{plotTSPsoln}
\alias{plotTSPsoln}
\title{Plot the TSP solution}
\usage{
plotTSPsoln(tourData, strainName, solverName, colorvector)
}
\arguments{
\item{tourData}{- TSP solution object, as produced by TSPinterface.R/processConcordeOutput or TSPinterface.R/processLKHOutput}

\item{strainName}{- name of the strain, as a string}

\item{solverName}{- name of the solver (Concorde or LKH), as a string}

\item{colorvector}{(OPTIONAL) - vector of color values to be used in the plot}
}
\description{
Plot the TSP solution.
}

