% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processLKHOutput.R
\name{processLKHOutput}
\alias{processLKHOutput}
\title{Process LKH solution file.}
\usage{
processLKHOutput(filename, markerdata, rfmat)
}
\arguments{
\item{filename}{- path and name of LKH solution file, which is returned from the function callLKH (this file has a .LKH extension)}

\item{markerdata}{- main marker data file, produced by either the readfile.R/readFile function or the duplicatemarkers.R/removedups function}

\item{rfmat}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}
}
\value{
TSP solution object, containing:
     1. the index of the markers in the main marker data object
     2. marker name
     3. rf value between each marker and the next
     4. cumulative rf value
}
\description{
Process LKH solution file.
}

