% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfile.R
\name{readFile}
\alias{readFile}
\title{This function reads in a raw data file from disk.}
\usage{
readFile(filename,classificationFlag,transpose)
}
\arguments{
\item{filename}{- path and input filename as a string}

\item{classificationFlag}{- flag to indicate whether the 2nd column in the data file is the classification column, which will be discarded.  Default is TRUE.}

\item{transpose}{- flag to indicate whether or not the data needs to be transposed.  Default is TRUE.}
}
\value{
main marker data object
}
\description{
This function reads in a raw data file from disk. This is the only place where data are brought into the package, all calculations are done from these data.
File format:
   1. Tab-separated or comma-separated file
   2. Markers are arranged in rows
   3. One header row which can be discarded
   4. One header column with the marker names as strings
   5. OPTIONAL: A second column which can also be discarded (this column typically has the string "(a,b)" in each cell).  If this column is not present, set the classificationFlag parameter to FALSE.
}

