% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangeGroup.R
\name{rearrangeGroup}
\alias{rearrangeGroup}
\title{Merge two clusters.}
\usage{
rearrangeGroup(group1Index, group2Index, cutlist, pathData)
}
\arguments{
\item{group1Index}{- index of the first group}

\item{group2Index}{- index of the second group}

\item{cutlist}{- cutlist array, produced by cutpoints.R/findcuts function}

\item{pathData}{- TSP solution object, produced by TSPinterface.R/processConcordeOutput or TSPinterface.R/processLKHOutput}
}
\value{
List of the new cutpointArray and TSP solution, after the two groups have been merged.
   Get the return values like this:
   result = rearrangeGroup(1, 2, cutpointArray, TSPsolution)
   newCutpointArray = result$v1
   newTSPsolution = result$v2
}
\description{
Merge two clusters.  This is done after finding the results of matchGroups and knowing which two groups are actually part of the same chromosome.=
}

