% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedups.R
\name{removedups}
\alias{removedups}
\title{This function removes duplicate markers from the main marker data object.}
\usage{
removedups(markerdata, duplist)
}
\arguments{
\item{markerdata}{- main marker data file, produced by the readfile.R/readFile function.}

\item{duplist}{- list of duplicate markers, produced by the duplicatemarkers.R/finddups function.}
}
\value{
main marker data file with duplicate markers removed
}
\description{
This function removes duplicate markers from the main marker data object.
}

