% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMarkers.R
\name{removeMarkers}
\alias{removeMarkers}
\title{Remove redundant markers.}
\usage{
removeMarkers(TSPsolution, redundantMarkerList, rfmat)
}
\arguments{
\item{TSPsolution}{- TSP solution object, produced by the TSPinterface.R/processConcordeOutput or TSPinterface.R/processLKHOutput functions}

\item{redundantMarkerList}{- list of redundant markers}

\item{rfmat}{- rf matrix}
}
\value{
TSP solution with redundant markers removed.
}
\description{
This function removes redundant markers from a TSP solution list.
}

