% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderMarkers.R
\name{reorderMarkers}
\alias{reorderMarkers}
\title{Reorder raw marker data}
\usage{
reorderMarkers(rawdata, distancedf)
}
\arguments{
\item{rawdata}{markerdata used to create map}

\item{distancedf}{dataframe containing markernames (markername) distances (`Haldane (cM)`) and groups (groupname) of markers}
}
\value{
data.frame of markers properly ordered with linkage group and position
}
\description{
Reorders raw markerdata and adds linkage group and marker position based on TSPmap results, if written as .csv can be imported to R/qtl by qtl:read.cross()
}

