% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseCluster.R
\name{reverseCluster}
\alias{reverseCluster}
\title{Reverse the order of markers in a group within a cluster.}
\usage{
reverseCluster(cluster, rfmat)
}
\arguments{
\item{cluster}{cliuster object}

\item{rfmat}{recombination frequency matrix, produced by the computeRFmat function}
}
\value{
cluster that is identical to cluster object, but with the markers of the selected  in reverse order.
   NOTE: this function updates the rf values and cumulative rf values, but does not update genetic distance values (if present).
}
\description{
Reverse the order of markers in a group within a  cluster
}

