% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGroup.R
\name{selectGroup}
\alias{selectGroup}
\title{Return the list of marker indices that comprise the given group.}
\usage{
selectGroup(groupIndex, markerData, cutList)
}
\arguments{
\item{groupIndex}{- index of the group to select}

\item{markerData}{- main marker data file, produced by either the readfile.R/readFile function or the duplicatemarkers.R/removedups function}

\item{cutList}{- cutpoint array, produced by cutpoints.R/findcuts function}
}
\value{
List of indices of the markers in the given group.
}
\description{
Return the list of marker indices that comprise the given group.
   This is an internal method that is used in the matchGroups function.
   GROUP SHOULD BE SHIFTED TO BEGIN AT A CUTPOINT BEFORE CALLING THIS???
}

