% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftTSPsoln.R
\name{shiftTSPsoln}
\alias{shiftTSPsoln}
\title{Shift a TSP solution array when the endpoints don't line up with the beginning of the array.}
\usage{
shiftTSPsoln(TSPsoln, startIndex)
}
\arguments{
\item{TSPsoln}{- array of markers of a TSP solution}

\item{startIndex}{- StartIndex is the index of the first cutpoint, which we will shift so that it is in position 1.  If startIndex is not specified, the default value of 0 will cause the method to shift the solution array so that it ends at the largest rf value (since this is probably the best cutpoint).}
}
\value{
Array of markers in TSP solution order, but shifted to startIndex.
}
\description{
This function returns a list of the smallest edge weight between a given cluster and all other clusters.  This is used to make the decision to merge clusters when the smallest edge weight is below a given threshold.  You will want to re-run the findCuts function after running this method.
}

