% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortRFmatrix.R
\name{sortRFmatrix}
\alias{sortRFmatrix}
\title{This function sorts the rfmatrix so that the markers are in the same order as the TSP tour.}
\usage{
sortRFmatrix(rfmatrix, pathData, saveFlag)
}
\arguments{
\item{rfmatrix}{- matrix of recombination frequency values produced by the rfmatrix.R/computeRFmat function}

\item{pathData}{- TSP tour and weights, produced by the TSPinterface.R/processConcordeOutput function}

\item{saveFlag}{- if saveFlag == 1, save sorted matrix to file}
}
\value{
matrix of recombination frequency values sorted in order of TSP tour
}
\description{
This function sorts the rfmatrix so that the markers are in the same order as the TSP tour.
   The resultant matrix of this function is used in the cutpoints.R/matchGroups function.
   NOTE: The resultant matrix of this function cannot be sent to the writeRFmat function because the columns/rows no longer match up with the main markerdata object.
}

