% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSPmap2distancedf.R
\name{TSPmap2distancedf}
\alias{TSPmap2distancedf}
\title{convert final tspmap to distance data.frame}
\usage{
TSPmap2distancedf(finalMap)
}
\arguments{
\item{finalMap}{final map list object created with TSPmap after grouping and ordering markers}
}
\value{
dataframe containing columns: index, markername, rf, cumulative rf, Haldane (cM), group, groupname
}
\description{
converts final map created by concorde to a dataframe including haldane distances and groupname
}

