% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeOutput.R
\name{writeOutput}
\alias{writeOutput}
\title{Write the TSP solution object to a comma-separated value file.}
\usage{
writeOutput(filename, TSPsol)
}
\arguments{
\item{filename}{- path and output filename as a string}

\item{TSPsol}{- TSP solution object, as produced by TSPinterface.R/processConcordeOutput or TSPinterface.R/processLKHOutput.}
}
\value{
None
}
\description{
Write the TSP solution object to a comma-separated value file.
}

