% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRFmat.R
\name{writeRFmat}
\alias{writeRFmat}
\title{writes the recombination frequency matrix to a file}
\usage{
writeRFmat(filename, markerdata, rfmatrix)
}
\arguments{
\item{filename}{- path and output filename as a string}

\item{markerdata}{- main marker data file, produced by either the readfile.R/readFile function or the duplicatemarkers.R/removedups function}

\item{rfmatrix}{- recombination frequency matrix, produced by the rfmatrix.R/computeRFmat function}
}
\value{
Rf matrix with marker names in the header row/column.
}
\description{
This function writes the recombination frequency matrix to a file, along with the marker names in both the header row and header column.
File format:
   1. Comma-separated file
}

