% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSortedRFmat.R
\name{writeSortedRFmat}
\alias{writeSortedRFmat}
\title{Write the sorted RF matrix to file.}
\usage{
writeSortedRFmat(filename, pathData, sortedrfmatrix)
}
\arguments{
\item{filename}{- filename to create}

\item{pathData}{- TSP solution with marker names}

\item{sortedrfmatrix}{- rf matrix in sorted order}
}
\value{
matrix of recombination frequency values with marker names added
}
\description{
This function saves a sorted RF matrix to file, along with the marker names.
}

